Create README.md
would create a new social media platform that would allow people to experience whatever the person on the other end wants to share.

For example, if someone sent you a video of their trip to the Great Wall of China, the video would be projected in 3D in front of you, so the viewer could feel like they were there, too. The app would immerse the viewer in a 360 degree hologram, and it would be called Hologram Connection.

The functions on the app would also be useful to companies that want to promote their business, and allow people to virtually try out their products. Seeing how quickly our world’s technology is advancing, I believe it is possible for this platform to be created.